correct data types in HungerOverhaul.cfg
When I was auditing the differences between the default Hunger Overhaul config and the current one, I noticed the data types of these config variables was different than in the defaults. (and the one in the defaults were more appropriately specific).

Dunno how the parser for this file works, or if it actually cares. (It seems like a weird thing to have as syntax in the config file and not in the code object it would load it in to.)